requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
Faker==18.6.2
prettytable==3.7.0